+++ Service/Services/ServiceInfoService.cs
@@ -1,6 +1,7 @@
 using Data.Interfaces;
 using Service.Dtos;
 using Service.Factories;
+using Service.Helpers;
 using Service.Interfaces;
 using Service.Models;
 
@@ -11,25 +12,27 @@ public class ServiceInfoService(IServiceInfoRepository serviceInfoRepository) : 
     private readonly IServiceInfoRepository _serviceInfoRepository = serviceInfoRepository;
     
     
-    public async Task<bool> CreateAsync(ServiceInfoDto? dto)
+    public async Task<IResult> CreateAsync(ServiceInfoDto? dto)
     {
-        if (dto is null || await _serviceInfoRepository.AlreadyExistsAsync(x => x.ServiceName == dto.ServiceName))
-            return false;
+        if (dto is null)
+            Result.BadRequest("Service Dto is null");
+                
+        if (await _serviceInfoRepository.AlreadyExistsAsync(x => x.ServiceName == dto.ServiceName))
+            return Result.AlreadyExists("Service name already exists");
         
         try
         {
             var entity = ServiceInfoFactory.Create(dto);
             await _serviceInfoRepository.CreateAsync(entity);
-            return true;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when creating Service: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was an error when creating Service: {e.Message}");
         }
     }
 
-    public async Task<IEnumerable<ServiceInfo>> GetAllAsync()
+    public async Task<IResult<IEnumerable<ServiceInfo>>> GetAllAsync()
     {
         var services = new List<ServiceInfo>();
         
@@ -37,57 +40,57 @@ public class ServiceInfoService(IServiceInfoRepository serviceInfoRepository) : 
         {
             var entities = await _serviceInfoRepository.GetAllAsync();
             services.AddRange(entities.Select(entity => ServiceInfoFactory.Create(entity)));
-            return services;
+            return Result<IEnumerable<ServiceInfo>>.Ok(services);
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when getting all Services: {e.Message}");
-            return [];
+            return Result<IEnumerable<ServiceInfo>>.ExceptionError($"There was an error when getting all Services: {e.Message}");
         }
     }
 
-    public async Task<ServiceInfo> GetByIdAsync(int id)
+    public async Task<IResult<ServiceInfo>> GetByIdAsync(int id)
     {
         try
         {
             var entity = await _serviceInfoRepository.GetAsync(x => x.Id == id);
-            return ServiceInfoFactory.Create(entity);
+            return Result<ServiceInfo>.Ok(ServiceInfoFactory.Create(entity));
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when getting Service: {e.Message}");
-            return null!;
+            return Result<ServiceInfo>.ExceptionError($"There was an error when getting Service: {e.Message}");
         }
     }
 
-    public async Task<bool> UpdateAsync(ServiceInfo? model)
+    public async Task<IResult> UpdateAsync(ServiceInfo? model)
     {
-        if (model is null) return false;
+        if (model is null) 
+            return Result.BadRequest("Service Model is null");
 
         try
         {
             var entity = ServiceInfoFactory.Create(model);
             await _serviceInfoRepository.UpdateAsync( (x => x.Id == model.Id), entity);
-            return true;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"there was error when updating The Service: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was an error when updating the Service: {e.Message}");
         }
     }
 
-    public async Task<bool> DeleteAsync(int id)
+    public async Task<IResult> DeleteAsync(int id)
     {
+        if (await _serviceInfoRepository.AlreadyExistsAsync(x => x.Id == id) == false)
+            return Result.BadRequest("Service Id Does Not Exists");
+        
         try
         {
             var result = await _serviceInfoRepository.DeleteAsync(x => x.Id == id);
-            return result;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"there was error when deleting service: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was an error when deleting service: {e.Message}");
         }
     }
 }
