+++ Service/Services/StatusInfoService.cs
@@ -1,6 +1,7 @@
 using Data.Interfaces;
 using Service.Dtos;
 using Service.Factories;
+using Service.Helpers;
 using Service.Interfaces;
 using Service.Models;
 
@@ -12,25 +13,28 @@ public class StatusInfoService(IStatusInfoRepository statusInfoRepository) : ISt
     
     
     
-    public async Task<bool> CreateAsync(StatusInfoDto? dto)
+    public async Task<IResult> CreateAsync(StatusInfoDto? dto)
     {
-        if (dto is null || await _statusInfoRepository.AlreadyExistsAsync(x => x.StatusName == dto.StatusName))
-            return false;
+        if (dto is null)
+            return Result.BadRequest("StatusInfoDto is null");
+                
+        if (await _statusInfoRepository.AlreadyExistsAsync(x => x.StatusName == dto.StatusName))
+            return Result.AlreadyExists("The status name already exists");
+        
         
         try
         {
             var entity = StatusInfoFactory.Create(dto);
             await _statusInfoRepository.CreateAsync(entity);
-            return true;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when creating StatusInfo: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was error when creating StatusInfo: {e.Message}");
         }
     }
 
-    public async Task<IEnumerable<StatusInfo>> GetAllAsync()
+    public async Task<IResult<IEnumerable<StatusInfo>>> GetAllAsync()
     {
         var statuses = new List<StatusInfo>();
         
@@ -38,57 +42,58 @@ public class StatusInfoService(IStatusInfoRepository statusInfoRepository) : ISt
         {
             var entities = await _statusInfoRepository.GetAllAsync();
             statuses.AddRange(entities.Select(entity => StatusInfoFactory.Create(entity)));
-            return statuses;
+            return Result<IEnumerable<StatusInfo>>.Ok(statuses);
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when getting all Statuses: {e.Message}");
-            return [];
+            return Result<IEnumerable<StatusInfo>>.ExceptionError($"There was error when getting all Statuses: {e.Message}");
         }
     }
 
-    public async Task<StatusInfo> GetByIdAsync(int id)
+    public async Task<IResult<StatusInfo>> GetByIdAsync(int id)
     {
         try
         {
             var entity = await _statusInfoRepository.GetAsync(x => x.Id == id);
-            return StatusInfoFactory.Create(entity);
+            //return Result<StatusInfo>.Ok(StatusInfoFactory.Create(entity));
+            return Result<StatusInfo>.Ok(StatusInfoFactory.Create(entity));
         }
         catch (Exception e)
         {
-            Console.WriteLine($"There was error when getting Status: {e.Message}");
-            return null!;
+            return Result<StatusInfo>.ExceptionError($"There was error when getting Status: {e.Message}");
         }
     }
 
-    public async Task<bool> UpdateAsync(StatusInfo? model)
+    public async Task<IResult> UpdateAsync(StatusInfo? model)
     {
-        if (model is null) return false;
+        if (model is null) 
+            return Result.BadRequest("StatusInfo is null");
 
         try
         {
             var entity = StatusInfoFactory.Create(model);
             await _statusInfoRepository.UpdateAsync( (x => x.Id == model.Id), entity);
-            return true;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"there was error when updating The Status: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was error when updating the Status: {e.Message}");
         }
     }
 
-    public async Task<bool> DeleteAsync(int id)
+    public async Task<IResult> DeleteAsync(int id)
     {
+        if (await _statusInfoRepository.AlreadyExistsAsync(x => x.Id == id) == false)
+            return Result.BadRequest("Role Id Does Not Exists");
+        
         try
         {
             var result = await _statusInfoRepository.DeleteAsync(x => x.Id == id);
-            return result;
+            return Result.Ok();
         }
         catch (Exception e)
         {
-            Console.WriteLine($"there was error when deleting Status: {e.Message}");
-            return false;
+            return Result.ExceptionError($"There was error when deleting Status: {e.Message}");
         }
     }
 }
